--- psd-web/app/controllers/investigations/view_meetings_controller.rb
@@ -1,9 +1,0 @@
-class Investigations::ViewMeetingsController < ApplicationController
-  def show
-    @investigation = Investigation.find_by!(pretty_id: params[:investigation_pretty_id])
-    authorize @investigation, :view_protected_details?
-    @meeting = @investigation.meetings.find(params[:id]).decorate
-
-    render "investigations/meetings/show"
-  end
-end
