+++ psd-web/app/views/investigations/record_corrective_actions/details.html.slim
@@ -12,7 +12,7 @@
       span.govuk-caption-l = @investigation.pretty_description
       h1.govuk-heading-l
         = page_title
-      = render "investigations/corrective_actions/form",
+      = render "investigations/record_corrective_actions/form",
                form: form,
                corrective_action: @corrective_action,
                investigation: @investigation,
