+++ psd-web/app/views/investigations/ts_investigations/corrective_action.html.slim
@@ -6,7 +6,7 @@
         hide_break: false,
         further_key: :further_corrective_action do |form|
 
-  = render "investigations/corrective_actions/form",
+  = render "investigations/record_corrective_actions/form",
            form: form,
            corrective_action: @corrective_action,
            investigation: @investigation,
